Finish: moving edge apps to separate mods
