+++ src/commands/edge_app/app.rs
@@ -1,10 +1,9 @@
-
 use crate::commands;
+use crate::commands::edge_app::instance_manifest::InstanceManifest;
 use crate::commands::edge_app::manifest::{EdgeAppManifest, Entrypoint};
 use crate::commands::edge_app::setting::{deserialize_settings_from_array, Setting, SettingType};
-use crate::commands::edge_app::instance_manifest::{InstanceManifest, INSTANCE_MANIFEST_VERSION};
-use crate::commands::{CommandError, EdgeAppInstances, EdgeAppSettings, EdgeApps};
 use crate::commands::edge_app::EdgeAppCommand;
+use crate::commands::{CommandError, EdgeApps};
 use indicatif::ProgressBar;
 use log::debug;
 use std::collections::HashMap;
@@ -13,7 +12,7 @@ use std::{io, str, thread};
 use reqwest::header::HeaderMap;
 use reqwest::StatusCode;
 use serde::{Deserialize, Serialize};
-use serde_json::{json, Value};
+use serde_json::json;
 use serde_yaml;
 use std::fs;
 use std::fs::File;
@@ -29,13 +28,10 @@ use crate::commands::edge_app::utils::{
     ensure_edge_app_has_all_necessary_files, generate_file_tree, FileChanges, SettingChanges,
 };
 
-use crate::commands::edge_app::server::{run_server, Metadata, MOCK_DATA_FILENAME};
 use crate::commands::edge_app::utils::transform_edge_app_path_to_manifest;
 
 use crate::commands::edge_app::manifest::{EntrypointType, MANIFEST_VERSION};
 use crate::commands::edge_app::utils::transform_instance_path_to_instance_manifest;
-
-
 
 #[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
 pub struct AssetSignature {
@@ -130,7 +126,8 @@ impl EdgeAppCommand {
 
         EdgeAppManifest::save_to_file(&manifest, path)?;
 
-        let index_html_template = include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/data/index.html"));
+        let index_html_template =
+            include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/data/index.html"));
         let index_html_file = File::create(&index_html_path)?;
         write!(&index_html_file, "{index_html_template}")?;
 
@@ -481,207 +478,6 @@ impl EdgeAppCommand {
     }
 }
 
-// Edge app settings commands
-impl EdgeAppCommand {
-    pub fn list_settings(&self, installation_id: &str) -> Result<EdgeAppSettings, CommandError> {
-        let app_id = self.get_app_id_by_installation(installation_id)?;
-
-        let app_settings: Vec<HashMap<String, serde_json::Value>> = serde_json::from_value(commands::get(&self.authentication,
-                                                                                                             &format!("v4.1/edge-apps/settings?select=name,type,default_value,optional,title,help_text,edge_app_setting_values(value)&app_id=eq.{}&order=name.asc",
-                                                                                                                      app_id,
-                                                                                                             ))?)?;
-
-        Ok(EdgeAppSettings::new(serde_json::to_value(app_settings)?))
-    }
-
-    pub fn set_setting(
-        &self,
-        path: Option<String>,
-        setting_key: &str,
-        setting_value: &str,
-    ) -> Result<(), CommandError> {
-        let installation_id = match self.get_installation_id(path.clone()) {
-            Ok(id) => Some(id),
-            Err(_) => None,
-        };
-        let app_id: String = match self.get_app_id(path.clone()) {
-            Ok(id) => id,
-            Err(_) => return Err(CommandError::MissingAppId),
-        };
-
-        let _is_setting_global = self.is_setting_global(&app_id, setting_key)?;
-
-        #[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-        struct SettingValue {
-            name: String,
-            #[serde(rename = "type")]
-            type_field: String,
-            edge_app_setting_values: Vec<HashMap<String, String>>,
-        }
-
-        let setting_url: String;
-        let settings_values_payload: Value;
-        let settings_values_patch_url: String;
-
-        if _is_setting_global {
-            setting_url = format!(
-                "v4.1/edge-apps/settings?select=name,type,edge_app_setting_values(value)&app_id=eq.{}&edge_app_setting_values.app_id=eq.{}&name=eq.{}",
-                app_id, app_id, setting_key,
-            );
-            settings_values_payload = json!(
-                {
-                    "app_id": app_id,
-                    "name": setting_key,
-                    "value": setting_value,
-                }
-            );
-            settings_values_patch_url = format!(
-                "v4.1/edge-apps/settings/values?app_id=eq.{}&name=eq.{}",
-                app_id, setting_key,
-            );
-        } else {
-            let actual_installation_id = match installation_id {
-                Some(id) => id,
-                None => return Err(CommandError::MissingInstallationId),
-            };
-
-            setting_url = format!(
-                "v4.1/edge-apps/settings?select=name,type,edge_app_setting_values(value)&edge_app_setting_values.installation_id=eq.{}&name=eq.{}&app_id=eq.{}",
-                actual_installation_id, setting_key, app_id
-            );
-            settings_values_payload = json!(
-                {
-                    "installation_id": actual_installation_id,
-                    "name": setting_key,
-                    "value": setting_value,
-                }
-            );
-            settings_values_patch_url = format!(
-                "v4.1/edge-apps/settings/values?installation_id=eq.{}&name=eq.{}",
-                actual_installation_id, setting_key,
-            );
-        }
-
-        let response = commands::get(&self.authentication, &setting_url)?;
-        let setting_values = serde_json::from_value::<Vec<SettingValue>>(response)?;
-
-        if setting_values.is_empty() {
-            commands::post(
-                &self.authentication,
-                "v4.1/edge-apps/settings/values",
-                &settings_values_payload,
-            )?;
-            return Ok(());
-        }
-        // we do know it is not empty - so it is safe to unwrap
-        let setting = setting_values.first().unwrap();
-
-        if setting.type_field == "secret" {
-            commands::post(
-                &self.authentication,
-                "v4.1/edge-apps/secrets/values",
-                &settings_values_payload,
-            )?;
-            return Ok(());
-        }
-
-        if setting.edge_app_setting_values.is_empty() {
-            commands::post(
-                &self.authentication,
-                "v4.1/edge-apps/settings/values",
-                &settings_values_payload,
-            )?;
-
-            return Ok(());
-        }
-
-        if setting.edge_app_setting_values.len() == 1
-            && setting.edge_app_setting_values[0].get("value").unwrap() == setting_value
-        {
-            println!("Setting value is already set to {}", setting_value);
-            return Ok(());
-        }
-        commands::patch(
-            &self.authentication,
-            &settings_values_patch_url,
-            &json!(
-                {
-                    "value": setting_value,
-                }
-            ),
-        )?;
-
-        Ok(())
-    }
-
-    pub fn run(&self, path: &Path, secrets: Vec<(String, String)>) -> Result<(), anyhow::Error> {
-        let address_shared = Arc::new(Mutex::new(None));
-        let address_clone = address_shared.clone();
-
-        let runtime = tokio::runtime::Runtime::new().unwrap();
-        let path = path.to_path_buf();
-        runtime.block_on(async {
-            tokio::spawn(async move {
-                let address = run_server(path.as_path(), secrets).await.unwrap();
-                let mut locked_address = address_clone.lock().unwrap();
-                *locked_address = Some(address);
-            })
-            .await
-            .unwrap();
-
-            while address_shared.lock().unwrap().is_none() {
-                tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-            }
-
-            println!(
-                "Edge App emulator is running at {}/index.html",
-                address_shared.lock().unwrap().as_ref().unwrap()
-            );
-
-            if let Err(e) = self.open_browser(&format!(
-                "{}/index.html",
-                address_shared.lock().unwrap().as_ref().unwrap()
-            )) {
-                eprintln!("{}", e);
-            }
-
-            loop {
-                tokio::time::sleep(std::time::Duration::from_secs(3600)).await;
-            }
-        });
-
-        Ok(())
-    }
-
-    fn open_browser(&self, address: &str) -> Result<(), CommandError> {
-        let command = match std::env::consts::OS {
-            "macos" => "open",
-            "windows" => "start",
-            "linux" => "xdg-open",
-            _ => {
-                return Err(CommandError::OpenBrowserError(
-                    "Unsupported OS to open browser".to_string(),
-                ))
-            }
-        };
-
-        let output = std::process::Command::new(command)
-            .arg(address)
-            .output()
-            .expect("Failed to open browser");
-
-        if !output.status.success() {
-            return Err(CommandError::OpenBrowserError(format!(
-                "Failed to open browser: {}",
-                str::from_utf8(&output.stderr).unwrap()
-            )));
-        }
-
-        Ok(())
-    }
-}
-
-
 impl EdgeAppCommand {
     pub fn get_app_name(&self, app_id: &str) -> Result<String, CommandError> {
         let response = commands::get(
@@ -709,47 +505,6 @@ impl EdgeAppCommand {
         manifest.id = None;
         EdgeAppManifest::save_to_file(&manifest, PathBuf::from(path).as_path())?;
 
-        Ok(())
-    }
-
-    pub fn generate_mock_data(&self, path: &Path) -> Result<(), CommandError> {
-        let data = fs::read_to_string(path)?;
-        let manifest: EdgeAppManifest = serde_yaml::from_str(&data)?;
-        let edge_app_dir = path.parent().ok_or(CommandError::MissingField)?;
-
-        if edge_app_dir.join(MOCK_DATA_FILENAME).exists() {
-            println!("Mock data for Edge App emulator already exists.");
-            return Ok(());
-        }
-
-        let default_metadata = Metadata::default();
-
-        let mut settings: HashMap<String, serde_yaml::Value> = HashMap::new();
-        for setting in &manifest.settings {
-            if setting.type_ != SettingType::Secret {
-                let settings_default_value = match setting.default_value {
-                    Some(ref default_value) => default_value.clone(),
-                    None => "".to_owned(),
-                };
-                settings.insert(
-                    setting.name.clone(),
-                    serde_yaml::Value::String(settings_default_value),
-                );
-            }
-        }
-
-        let mut mock_data: HashMap<String, serde_yaml::Value> = HashMap::new();
-        mock_data.insert(
-            "metadata".to_string(),
-            serde_yaml::to_value(default_metadata)?,
-        );
-        mock_data.insert("settings".to_string(), serde_yaml::to_value(settings)?);
-
-        let mock_data_yaml = serde_yaml::to_string(&mock_data)?;
-
-        fs::write(edge_app_dir.join(MOCK_DATA_FILENAME), mock_data_yaml)?;
-
-        println!("Mock data for Edge App emulator was generated.");
         Ok(())
     }
 
@@ -889,70 +644,6 @@ impl EdgeAppCommand {
         Ok(())
     }
 
-    fn create_setting(&self, app_id: String, setting: &Setting) -> Result<(), CommandError> {
-        let value = serde_json::to_value(setting)?;
-        let mut payload = serde_json::from_value::<HashMap<String, serde_json::Value>>(value)?;
-        payload.insert("app_id".to_owned(), json!(app_id));
-        payload.insert("name".to_owned(), json!(setting.name));
-
-        debug!("Creating setting: {:?}", &payload);
-
-        let response = commands::post(&self.authentication, "v4.1/edge-apps/settings", &payload);
-        if response.is_err() {
-            let c = commands::get(
-                &self.authentication,
-                &format!("v4.1/edge-apps/settings?app_id=eq.{}", app_id),
-            )?;
-            debug!("Existing settings: {:?}", c);
-            return Err(CommandError::NoChangesToUpload("".to_owned()));
-        }
-
-        Ok(())
-    }
-
-    fn update_setting(&self, app_id: String, setting: &Setting) -> Result<(), CommandError> {
-        let value = serde_json::to_value(setting)?;
-        let mut payload = serde_json::from_value::<HashMap<String, serde_json::Value>>(value)?;
-        payload.insert("name".to_owned(), json!(setting.name));
-
-        debug!("Updating setting: {:?}", &payload);
-
-        let response = commands::patch(
-            &self.authentication,
-            &format!(
-                "v4.1/edge-apps/settings?app_id=eq.{id}&name=eq.{name}",
-                id = app_id,
-                name = setting.name
-            ),
-            &payload,
-        );
-
-        if let Err(error) = response {
-            debug!("Failed to update setting: {}", setting.name);
-            return Err(error);
-        }
-
-        Ok(())
-    }
-
-    fn delete_setting(&self, app_id: String, setting: &Setting) -> Result<(), CommandError> {
-        let response = commands::delete(
-            &self.authentication,
-            &format!(
-                "v4.1/edge-apps/settings?app_id=eq.{id}&name=eq.{name}",
-                id = app_id,
-                name = setting.name
-            ),
-        );
-
-        if let Err(error) = response {
-            debug!("Failed to delete setting: {}", setting.name);
-            return Err(error);
-        }
-
-        Ok(())
-    }
-
     fn try_delete_setting(
         &self,
         app_id: String,
@@ -1073,40 +764,6 @@ impl EdgeAppCommand {
         Ok(())
     }
 
-    pub fn install_edge_app(
-        &self,
-        app_id: &str,
-        name: &str,
-        entrypoint: Option<String>,
-    ) -> Result<String, CommandError> {
-        let mut payload = json!({
-            "app_id": app_id,
-            "name": name,
-        });
-
-        if let Some(_entrypoint) = entrypoint {
-            payload["entrypoint"] = json!(_entrypoint);
-        }
-
-        let response = commands::post(
-            &self.authentication,
-            "v4.1/edge-apps/installations?select=id",
-            &payload,
-        )?;
-
-        #[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-        struct Installation {
-            id: String,
-        }
-
-        let installation = serde_json::from_value::<Vec<Installation>>(response)?;
-        if installation.is_empty() {
-            return Err(CommandError::MissingField);
-        }
-
-        Ok(installation[0].id.clone())
-    }
-
     fn publish(&self, app_id: &str, revision: u32) -> Result<(), CommandError> {
         commands::patch(
             &self.authentication,
@@ -1120,7 +777,7 @@ impl EdgeAppCommand {
         Ok(())
     }
 
-    fn is_setting_global(&self, app_id: &str, setting_key: &str) -> Result<bool, CommandError> {
+    pub fn is_setting_global(&self, app_id: &str, setting_key: &str) -> Result<bool, CommandError> {
         let response = commands::get(
             &self.authentication,
             &format!(
@@ -1213,17 +870,16 @@ impl EdgeAppCommand {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::authentication::Config;
     use std::env;
-    use tempfile::TempDir;
 
     use commands::edge_app::manifest::MANIFEST_VERSION;
     use httpmock::Method::{DELETE, GET, PATCH, POST};
 
-    use crate::authentication::Authentication;
-    use crate::commands::edge_app::server::MOCK_DATA_FILENAME;
+    use crate::commands::edge_app::test_utils::tests::{
+        create_edge_app_manifest_for_test, create_instance_manifest_for_test,
+        prepare_edge_apps_test,
+    };
     use crate::commands::edge_app::utils::EdgeAppFile;
-    use crate::commands::edge_app::test_utils::tests::{prepare_edge_apps_test, create_edge_app_manifest_for_test, create_instance_manifest_for_test};
     use tempfile::tempdir;
 
     #[test]
@@ -1289,7 +945,10 @@ mod tests {
         );
 
         let data_index_html = fs::read_to_string(tmp_dir.path().join("index.html")).unwrap();
-        assert_eq!(data_index_html, include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/data/index.html")));
+        assert_eq!(
+            data_index_html,
+            include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/data/index.html"))
+        );
 
         assert!(result.is_ok());
     }
@@ -1447,721 +1106,6 @@ mod tests {
 
         let result = command.list();
         edge_apps_mock.assert();
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_list_settings_should_send_correct_request() {
-        let (_tmp_dir, command, mock_server, _manifest, instance_manifest) =
-            prepare_edge_apps_test(false, true);
-
-        let installations_get_mock = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/installations")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("select", "app_id")
-                .query_param("id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEB");
-            then.status(200).json_body(json!([
-                {
-                    "app_id": "02H2QZ6Z8WXWNDC0KQ198XCZEW"
-                }
-            ]));
-        });
-
-        // &format!("v4.1/edge-apps/settings?select=name,type,default_value,optional,title,help_text,edge_app_setting_values(value)&app_id=eq.{}&order=name.asc",
-        let settings_mock = mock_server.mock(|when, then| {
-                when.method(GET)
-                    .path("/v4.1/edge-apps/settings")
-                    .header("Authorization", "Token token")
-                    .header(
-                        "user-agent",
-                        format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                    )
-                    .query_param("select", "name,type,default_value,optional,title,help_text,edge_app_setting_values(value)")
-                    .query_param("app_id", "eq.02H2QZ6Z8WXWNDC0KQ198XCZEW")
-                    .query_param("order", "name.asc");
-
-                then.status(200).json_body(json!([
-                    {
-                        "name": "Example setting1",
-                        "type": "string",
-                        "default_value": "stranger",
-                        "optional": true,
-                        "title": "Example title1",
-                        "help_text": "An example of a setting that is used in index.html",
-                        "edge_app_setting_values": [
-                            {
-                                "value": "stranger1"
-                            }
-                        ]
-                    },
-                    {
-                        "name": "Example setting2",
-                        "type": "string",
-                        "default_value": "stranger",
-                        "optional": true,
-                        "title": "Example title2",
-                        "help_text": "An example of a setting that is used in index.html",
-                        "edge_app_setting_values": [
-                            {
-                                "value": "stranger2"
-                            }
-                        ]
-                    },
-                    {
-                        "name": "Example setting3",
-                        "type": "string",
-                        "default_value": "stranger",
-                        "optional": true,
-                        "title": "Example title3",
-                        "help_text": "An example of a setting that is used in index.html",
-                        "edge_app_setting_values": []
-                    },
-                    {
-                        "name": "Example secret",
-                        "type": "secret",
-                        "default_value": "stranger",
-                        "optional": true,
-                        "title": "Example title4",
-                        "help_text": "An example of a secret that is used in index.html",
-                        "edge_app_setting_values": []
-                    }
-                ]));
-            });
-
-        let result = command.list_settings(&instance_manifest.unwrap().id.unwrap());
-
-        installations_get_mock.assert();
-        settings_mock.assert();
-
-        assert!(result.is_ok());
-        let settings = result.unwrap();
-        let settings_json: Value = serde_json::from_value(settings.value).unwrap();
-        assert_eq!(
-            settings_json,
-            json!([
-                {
-                    "name": "Example setting1",
-                    "type": "string",
-                    "default_value": "stranger",
-                    "optional": true,
-                    "title": "Example title1",
-                    "help_text": "An example of a setting that is used in index.html",
-                    "edge_app_setting_values": [
-                        {
-                            "value": "stranger1"
-                        }
-                    ]
-                },
-                {
-                    "name": "Example setting2",
-                    "type": "string",
-                    "default_value": "stranger",
-                    "optional": true,
-                    "title": "Example title2",
-                    "help_text": "An example of a setting that is used in index.html",
-                    "edge_app_setting_values": [
-                        {
-                            "value": "stranger2"
-                        }
-                    ]
-                },
-                {
-                    "name": "Example setting3",
-                    "type": "string",
-                    "default_value": "stranger",
-                    "optional": true,
-                    "title": "Example title3",
-                    "help_text": "An example of a setting that is used in index.html",
-                    "edge_app_setting_values": []
-                },
-                {
-                    "name": "Example secret",
-                    "type": "secret",
-                    "default_value": "stranger",
-                    "optional": true,
-                    "title": "Example title4",
-                    "help_text": "An example of a secret that is used in index.html",
-                    "edge_app_setting_values": []
-                }
-            ])
-        );
-    }
-
-    #[test]
-    fn test_set_setting_should_send_correct_request() {
-        let (tmp_dir, command, mock_server, _manifest, _instance_manifest) =
-            prepare_edge_apps_test(true, true);
-
-        let setting_get_is_global_mock = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("select", "is_global")
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW")
-                .query_param("name", "eq.best_setting");
-
-            then.status(200).json_body(json!([
-                {
-                    "is_global": false,
-                }
-            ]));
-        });
-
-        // "v4/edge-apps/settings/values?select=title&installation_id=eq.{}&title=eq.{}"
-        let setting_mock_get = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("name", "eq.best_setting")
-                .query_param("select", "name,type,edge_app_setting_values(value)")
-                .query_param(
-                    "edge_app_setting_values.installation_id",
-                    "eq.01H2QZ6Z8WXWNDC0KQ198XCZEB",
-                )
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW");
-            then.status(200).json_body(json!([]));
-        });
-
-        let setting_values_mock_post = mock_server.mock(|when, then| {
-            when.method(POST)
-                .path("/v4.1/edge-apps/settings/values")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .json_body(json!(
-                    {
-                        "name": "best_setting",
-                        "value": "best_value",
-                        "installation_id": "01H2QZ6Z8WXWNDC0KQ198XCZEB"
-                    }
-                ));
-            then.status(204).json_body(json!({}));
-        });
-
-        let result = command.set_setting(
-            Some(tmp_dir.path().to_str().unwrap().to_string()),
-            "best_setting",
-            "best_value",
-        );
-
-        setting_get_is_global_mock.assert();
-        setting_mock_get.assert();
-        setting_values_mock_post.assert();
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_set_setting_when_setting_value_exists_should_send_correct_update_request() {
-        let (tmp_dir, command, mock_server, _manifest, _instance_manifest) =
-            prepare_edge_apps_test(true, true);
-
-        let setting_get_is_global_mock = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("select", "is_global")
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW")
-                .query_param("name", "eq.best_setting");
-
-            then.status(200).json_body(json!([
-                {
-                    "is_global": false,
-                }
-            ]));
-        });
-
-        // "v4/edge-apps/settings/values?select=title&installation_id=eq.{}&title=eq.{}"
-        let setting_mock_get = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("name", "eq.best_setting")
-                .query_param("select", "name,type,edge_app_setting_values(value)")
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW")
-                .query_param(
-                    "edge_app_setting_values.installation_id",
-                    "eq.01H2QZ6Z8WXWNDC0KQ198XCZEB",
-                );
-            then.status(200).json_body(json!([
-                {
-                    "name": "best_setting",
-                    "type": "string",
-                    "edge_app_setting_values": [
-                        {
-                            "value": "best_value"
-                        }
-                    ]
-
-                }
-            ]));
-        });
-
-        let setting_values_mock_patch = mock_server.mock(|when, then| {
-            when.method(PATCH)
-                .path("/v4.1/edge-apps/settings/values")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("name", "eq.best_setting")
-                .query_param("installation_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEB")
-                .json_body(json!(
-                    {
-                        "value": "best_value1",
-                    }
-                ));
-            then.status(200).json_body(json!({}));
-        });
-
-        let result = command.set_setting(
-            Some(tmp_dir.path().to_str().unwrap().to_string()),
-            "best_setting",
-            "best_value1",
-        );
-
-        setting_get_is_global_mock.assert();
-        setting_mock_get.assert();
-        setting_values_mock_patch.assert();
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_set_global_setting_when_setting_value_exists_should_send_correct_update_request() {
-        let (temp_dir, command, mock_server, _manifest, _instance_manifest) =
-            prepare_edge_apps_test(true, true);
-
-        let setting_is_global_get_mock = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("select", "is_global")
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW")
-                .query_param("name", "eq.best_setting");
-
-            then.status(200).json_body(json!([
-                {
-                    "is_global": true,
-                }
-            ]));
-        });
-
-        // "v4.1/edge-apps/settings?select=name,type,edge_app_setting_values(value)&edge_app_setting_values.app_id=eq.{}&name=eq.{}",
-        let setting_mock_get = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("name", "eq.best_setting")
-                .query_param("select", "name,type,edge_app_setting_values(value)")
-                .query_param(
-                    "edge_app_setting_values.app_id",
-                    "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW",
-                )
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW");
-            then.status(200).json_body(json!([
-                {
-                    "name": "best_setting",
-                    "type": "string",
-                    "edge_app_setting_values": [
-                        {
-                            "value": "best_value"
-                        }
-                    ]
-                }
-            ]));
-        });
-
-        let setting_values_mock_patch = mock_server.mock(|when, then| {
-            when.method(PATCH)
-                .path("/v4.1/edge-apps/settings/values")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("name", "eq.best_setting")
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW")
-                .json_body(json!(
-                    {
-                        "value": "best_value1",
-                    }
-                ));
-            then.status(200).json_body(json!({}));
-        });
-
-        let result = command.set_setting(
-            Some(temp_dir.path().to_str().unwrap().to_string()),
-            "best_setting",
-            "best_value1",
-        );
-
-        setting_is_global_get_mock.assert();
-        setting_mock_get.assert();
-        setting_values_mock_patch.assert();
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_set_global_setting_when_setting_value_not_exists_should_send_correct_create_request() {
-        let (temp_dir, command, mock_server, _manifest, _instance_manifest) =
-            prepare_edge_apps_test(true, true);
-
-        let setting_is_global_get_mock = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("select", "is_global")
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW")
-                .query_param("name", "eq.best_setting");
-
-            then.status(200).json_body(json!([
-                {
-                    "is_global": true,
-                }
-            ]));
-        });
-
-        let setting_mock_get = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("name", "eq.best_setting")
-                .query_param("select", "name,type,edge_app_setting_values(value)")
-                .query_param(
-                    "edge_app_setting_values.app_id",
-                    "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW",
-                )
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW");
-            then.status(200).json_body(json!([]));
-        });
-
-        let setting_values_mock_post = mock_server.mock(|when, then| {
-            when.method(POST)
-                .path("/v4.1/edge-apps/settings/values")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .json_body(json!(
-                    {
-                        "value": "best_value1",
-                        "name": "best_setting",
-                        "app_id": "01H2QZ6Z8WXWNDC0KQ198XCZEW",
-                    }
-                ));
-            then.status(200).json_body(json!({}));
-        });
-
-        let result = command.set_setting(
-            Some(temp_dir.path().to_str().unwrap().to_string()),
-            "best_setting",
-            "best_value1",
-        );
-
-        setting_is_global_get_mock.assert();
-        setting_mock_get.assert();
-        setting_values_mock_post.assert();
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_set_setting_when_setting_doesnt_exist_should_fail() {
-        let (temp_dir, command, mock_server, _manifest, _instance_manifest) =
-            prepare_edge_apps_test(true, true);
-
-        let setting_get_mock = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("select", "is_global")
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW")
-                .query_param("name", "eq.best_setting");
-
-            then.status(200).json_body(json!([]));
-        });
-
-        let result = command.set_setting(
-            Some(temp_dir.path().to_str().unwrap().to_string()),
-            "best_setting",
-            "best_value1",
-        );
-
-        setting_get_mock.assert();
-        assert!(result.is_err());
-        let error = result.unwrap_err();
-        assert_eq!(
-            error.to_string(),
-            "Setting does not exist: best_setting.".to_string()
-        );
-    }
-
-    #[test]
-    fn test_set_setting_with_secret_should_send_correct_request() {
-        let (temp_dir, command, mock_server, _manifest, _instance_manifest) =
-            prepare_edge_apps_test(true, true);
-
-        let setting_is_global_get_mock = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("select", "is_global")
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW")
-                .query_param("name", "eq.best_secret_setting");
-
-            then.status(200).json_body(json!([
-                {
-                    "is_global": false,
-                }
-            ]));
-        });
-
-        let setting_mock_get = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("name", "eq.best_secret_setting")
-                .query_param("select", "name,type,edge_app_setting_values(value)")
-                .query_param(
-                    "edge_app_setting_values.installation_id",
-                    "eq.01H2QZ6Z8WXWNDC0KQ198XCZEB",
-                )
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW");
-            then.status(200).json_body(json!([
-                {
-                    "name": "best_secret_setting",
-                    "type": "secret",
-                    "edge_app_setting_values": []
-                }
-            ]));
-        });
-
-        // "v4/edge-apps/secrets/values"
-        let secrets_values_mock_post = mock_server.mock(|when, then| {
-            when.method(POST)
-                .path("/v4.1/edge-apps/secrets/values")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .json_body(json!(
-                    {
-                        "name": "best_secret_setting",
-                        "value": "best_secret_value",
-                        "installation_id": "01H2QZ6Z8WXWNDC0KQ198XCZEB"
-                    }
-                ));
-            then.status(204).json_body(json!({}));
-        });
-
-        let result = command.set_setting(
-            Some(temp_dir.path().to_str().unwrap().to_string()),
-            "best_secret_setting",
-            "best_secret_value",
-        );
-
-        setting_is_global_get_mock.assert();
-        setting_mock_get.assert();
-        secrets_values_mock_post.assert();
-        debug!("result: {:?}", result);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_set_global_secrets_should_send_correct_request() {
-        let (temp_dir, command, mock_server, _manifest, _instance_manifest) =
-            prepare_edge_apps_test(true, true);
-
-        let setting_is_global_get_mock = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("select", "is_global")
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW")
-                .query_param("name", "eq.best_secret_setting");
-
-            then.status(200).json_body(json!([
-                {
-                    "is_global": true,
-                }
-            ]));
-        });
-
-        let setting_mock_get = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("name", "eq.best_secret_setting")
-                .query_param("select", "name,type,edge_app_setting_values(value)")
-                .query_param(
-                    "edge_app_setting_values.app_id",
-                    "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW",
-                )
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW");
-            then.status(200).json_body(json!([
-                {
-                    "name": "best_secret_setting",
-                    "type": "secret",
-                    "edge_app_setting_values": []
-                }
-            ]));
-        });
-
-        // "v4/edge-apps/secrets/values"
-
-        let secrets_values_mock_post = mock_server.mock(|when, then| {
-            when.method(POST)
-                .path("/v4.1/edge-apps/secrets/values")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .json_body(json!(
-                    {
-                        "name": "best_secret_setting",
-                        "value": "best_secret_value",
-                        "app_id": "01H2QZ6Z8WXWNDC0KQ198XCZEW"
-                    }
-                ));
-            then.status(204).json_body(json!({}));
-        });
-
-        let result = command.set_setting(
-            Some(temp_dir.path().to_str().unwrap().to_string()),
-            "best_secret_setting",
-            "best_secret_value",
-        );
-
-        setting_is_global_get_mock.assert();
-        setting_mock_get.assert();
-        secrets_values_mock_post.assert();
-        debug!("result: {:?}", result);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_set_setting_when_value_has_not_changed_should_not_update_it() {
-        let (temp_dir, command, mock_server, _manifest, _instance_manifest) =
-            prepare_edge_apps_test(true, true);
-
-        let setting_get_is_global_mock = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("select", "is_global")
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW")
-                .query_param("name", "eq.best_setting");
-
-            then.status(200).json_body(json!([
-                {
-                    "is_global": false,
-                }
-            ]));
-        });
-
-        // "v4/edge-apps/settings/values?select=title&installation_id=eq.{}&title=eq.{}"
-        let setting_mock_get = mock_server.mock(|when, then| {
-            when.method(GET)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .query_param("name", "eq.best_setting")
-                .query_param("select", "name,type,edge_app_setting_values(value)")
-                .query_param("app_id", "eq.01H2QZ6Z8WXWNDC0KQ198XCZEW")
-                .query_param(
-                    "edge_app_setting_values.installation_id",
-                    "eq.01H2QZ6Z8WXWNDC0KQ198XCZEB",
-                );
-            then.status(200).json_body(json!([
-                {
-                    "name": "best_setting",
-                    "type": "string",
-                    "edge_app_setting_values": [
-                        {
-                            "value": "best_value"
-                        }
-                    ]
-
-                }
-            ]));
-        });
-
-        let result = command.set_setting(
-            Some(temp_dir.path().to_str().unwrap().to_string()),
-            "best_setting",
-            "best_value",
-        );
-
-        setting_get_is_global_mock.assert();
-        setting_mock_get.assert();
         assert!(result.is_ok());
     }
 
@@ -2699,96 +1643,6 @@ mod tests {
     }
 
     #[test]
-    fn test_generate_mock_data_creates_file_with_expected_content() {
-        let (_dir, command, _mock_server, _manifest, _instance_manifest) =
-            prepare_edge_apps_test(false, false);
-        let dir = tempdir().unwrap();
-        let file_path = dir.path().join("test_manifest.yml");
-
-        // The EdgeAppManifest structure from your example
-        let manifest = create_edge_app_manifest_for_test(vec![
-            Setting {
-                name: "asetting".to_string(),
-                type_: SettingType::String,
-                title: Some("atitle".to_string()),
-                optional: false,
-                default_value: Some("yes".to_string()),
-                is_global: false,
-                help_text: "help text".to_string(),
-            },
-            Setting {
-                name: "nsetting".to_string(),
-                type_: SettingType::String,
-                title: Some("ntitle".to_string()),
-                optional: false,
-                default_value: Some("".to_string()),
-                is_global: false,
-                help_text: "help text".to_string(),
-            },
-        ]);
-
-        EdgeAppManifest::save_to_file(&manifest, &file_path).unwrap();
-        command.generate_mock_data(&file_path).unwrap();
-
-        let mock_data_path = dir.path().join(MOCK_DATA_FILENAME);
-        assert!(mock_data_path.exists());
-
-        let _generated_content = fs::read_to_string(&mock_data_path).unwrap();
-        let _expected_content = r#"metadata:
-      coordinates:
-        - "37.3861"
-        - "-122.0839"
-      hostname: "srly-t6kb0ta1jrd9o0w"
-      location: "Code Cafe, Mountain View, California"
-      screen_name: "Code Cafe Display"
-      tags:
-        - "All Screens"
-    settings:
-      asetting: "yes"
-      nsetting: ""
-    "#;
-    }
-
-    #[test]
-    fn test_generate_mock_data_excludes_secret_settings() {
-        let dir = tempdir().unwrap();
-        let file_path = dir.path().join("test_manifest_with_varied_settings.yml");
-
-        let manifest = create_edge_app_manifest_for_test(vec![
-            Setting {
-                name: "excluded_setting".to_string(),
-                type_: SettingType::Secret,
-                title: Some("excluded title".to_string()),
-                optional: false,
-                default_value: None,
-                is_global: false,
-                help_text: "help text".to_string(),
-            },
-            Setting {
-                name: "included_setting".to_string(),
-                type_: SettingType::String,
-                title: Some("included title".to_string()),
-                optional: false,
-                default_value: Some("".to_string()),
-                is_global: false,
-                help_text: "help text".to_string(),
-            },
-        ]);
-
-        EdgeAppManifest::save_to_file(&manifest, &file_path).unwrap();
-        let config = Config::new("".to_owned());
-        let authentication = Authentication::new_with_config(config, "token");
-        let command = EdgeAppCommand::new(authentication);
-        command.generate_mock_data(&file_path).unwrap();
-
-        let mock_data_path = dir.path().join(MOCK_DATA_FILENAME);
-        let content = fs::read_to_string(mock_data_path).unwrap();
-
-        assert!(!content.contains("excluded_setting"));
-        assert!(content.contains("included_setting"));
-    }
-
-    #[test]
     fn test_ensure_assets_processing_finished_when_processing_failed_should_return_error() {
         let (temp_dir, command, mock_server, _manifest, _instance_manifest) =
             prepare_edge_apps_test(false, false);
@@ -2932,7 +1786,7 @@ mod tests {
     }
 
     #[test]
-    fn test_upload_without_app_id_should_fail() {
+    fn test_deploy_without_app_id_should_fail() {
         let (temp_dir, command, _mock_server, _manifest, _instance_manifest) =
             prepare_edge_apps_test(false, false);
 
@@ -3196,59 +2050,6 @@ mod tests {
         copy_assets_mock.assert();
 
         assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_create_is_global_setting_should_pass_is_global_property() {
-        let (_temp_dir, command, mock_server, _manifest, _instance_manifest) =
-            prepare_edge_apps_test(true, false);
-
-        //  v4/edge-apps/settings?app_id=eq.{}
-        let settings_mock_create = mock_server.mock(|when, then| {
-            when.method(POST)
-                .path("/v4.1/edge-apps/settings")
-                .header("Authorization", "Token token")
-                .header(
-                    "user-agent",
-                    format!("screenly-cli {}", env!("CARGO_PKG_VERSION")),
-                )
-                .json_body(json!({
-                    "name": "ssetting",
-                    "app_id": "01H2QZ6Z8WXWNDC0KQ198XCZEW",
-                    "type": "secret",
-                    "default_value": "",
-                    "title": "stitle",
-                    "optional": false,
-                    "help_text": "help text",
-                    "is_global": true
-                }));
-            then.status(201).json_body(json!(
-            [{
-                "name": "ssetting",
-                "app_id": "01H2QZ6Z8WXWNDC0KQ198XCZEW",
-                "type": "secret",
-                "default_value": "",
-                "title": "stitle",
-                "optional": false,
-                "help_text": "help text",
-                "is_global": true,
-            }]));
-        });
-
-        let setting = Setting {
-            name: "ssetting".to_string(),
-            type_: SettingType::Secret,
-            title: Some("stitle".to_string()),
-            optional: false,
-            default_value: Some("".to_string()),
-            is_global: true,
-            help_text: "help text".to_string(),
-        };
-        command
-            .create_setting("01H2QZ6Z8WXWNDC0KQ198XCZEW".to_string(), &setting)
-            .unwrap();
-
-        settings_mock_create.assert();
     }
 
     #[test]
